프리펩/hp 추가
canvas/player 프리펩수정
hpmanager/staminamanager 추가
