--- Assets/Scripts/UI/PlayerStatePanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3c4dfbd67007ca4a8705e6945e38959
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
