+++ Assets/TextMesh Pro/Resources/Fonts & Materials/BebasNeue-Regular SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ceb02fda6aab8a4abda60606fbf3d3d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
