+++ Assets/TextMesh Pro/Resources/Fonts & Materials/Digital Glitch Demo SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b759642029425fb4280b7815fb8589a0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
